UI updates

+++ client/src/components/Profile.js
@@ -5,6 +5,9 @@ import './Profile.css'
 import { Link } from 'react-router-dom';
 import { useNavigate } from 'react-router-dom';
 import checkmark from './images/yellow_checkmark.png'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faEdit, faEye, faTrashAlt } from '@fortawesome/free-solid-svg-icons';
+import Swal from 'sweetalert2';
 
 function Profile() {
     const navigate = useNavigate();
@@ -38,6 +41,7 @@ function Profile() {
     const [userPosts, setUserPosts] = useState([])
     const [userOrgs, setUserOrgs] = useState([])
     const [purdueEmail, setPurdueEmail] = useState(false)
+    const [dropdownVisible, setDropdownVisible] = useState({});
 
     useEffect(() => {
         async function fetchUser() {
@@ -73,26 +77,58 @@ function Profile() {
         navigate('/');
     };
 
-    const handleDeleteEvent = async (eventId) => {
-        try {
-            const response = await axios.delete(`http://localhost:8000/events/delete/${eventId}`);
-            alert("Event deleted")
-            console.log('Event deleted successfully:', response.data);
-          } catch (error) {
-            console.error('Error deleting event:', error);
-        }
+    const handleDeleteEvent = (eventId) => {
+        Swal.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!'
+        }).then(async (result) => {
+          if (result.isConfirmed) {
+            try {
+              const response = await axios.delete(`http://localhost:8000/events/delete/${eventId}`);
+              Swal.fire('Deleted!', 'Your event has been deleted.', 'success');
+              console.log('Event deleted successfully:', response.data);
+            } catch (error) {
+              Swal.fire('Error!', 'There was a problem deleting your event.', 'error');
+              console.error('Error deleting event:', error);
+            }
+          }
+        });
     };
 
-    const handleDeletePost = async (postId) => {
-        try {
-            const response = await axios.delete(`http://localhost:8000/posts/delete/${userId}/${postId}`);
-            console.log('Post deleted successfully:', response.data);
-          } catch (error) {
-            console.error('Error deleting event:', error);
-        }
+    const handleDeletePost = (postId) => {      
+        Swal.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!'
+        }).then(async (result) => {
+          if (result.isConfirmed) {
+            try {
+              const response = await axios.delete(`http://localhost:8000/posts/delete/${userId}/${postId}`);
+              Swal.fire('Deleted!', 'Your post has been deleted.', 'success');
+              console.log('Post deleted successfully:', response.data);
+            } catch (error) {
+              Swal.fire('Error!', 'There was a problem deleting your post.', 'error');
+              console.error('Error deleting post:', error);
+            }
+          }
+        });
+      };
+    
+    const toggleDropdown = (eventId) => {
+        setDropdownVisible(prevState => ({
+          ...prevState,
+          [eventId]: !prevState[eventId]
+        }));
     };
-      
-      
 
     const { name, bio, emailNotifs, createdDatetime, followingOrgs, prevInterestedEvents, posts } = user;
 
@@ -103,53 +139,56 @@ function Profile() {
                 {purdueEmail ? (<img className="verified-checkmark" src={checkmark} />)  : <></>}  
             </div>
             <div className='hosted-events'>
+                <h1>Hosted Events:</h1>
                 {userEvents ? (
                     userEvents.map(event => (
-                        <div className='user-event-row'>
-                            <div className='buttons' style={{ display: 'flex', alignItems: 'center', marginBottom: '10px' }}>
-                                <Link to={`/event/${event._id}`} style={{ marginRight: '10px' }}>
-                                    <div className={`event ${event.category}`} style={{ marginRight: '10px'}}>{event.title}</div>
-                                </Link>
-                                <Link to={`/edit-event/${event._id}`} style={{ marginRight: '10px' }}>
-                                    <button>EDIT</button>
-                                </Link>
-                                <button onClick={() => handleDeleteEvent(event._id)}>Delete</button>
-                            </div>
-                            {<div className='attendants'>
-                                {event.usersInterested.length ? (
-                                    <div>Users attending:</div>
-                                ) :(<div></div>)}
-                                {event.usersInterested.map(userInterested => (
-                                    <p>{userInterested.name}</p>
-                                ))}
+                        <div className='hosted-event'>
+                            <p>{event.title}</p>
+                            <div className='attendants'>
+                                <button onClick={() => toggleDropdown(event._id)}>
+                                    Users attending: {event.usersInterested.length}
+                                </button>
+                                <div className={`dropdown ${dropdownVisible[event._id] ? 'show' : ''}`}>
+                                    {event.usersInterested.map(userInterested => (
+                                        <p key={userInterested._id}>{userInterested.name}</p>
+                                    ))}
                                 </div>
-                            }
+                            </div>
+                            <div>
+                                <Link to={`/event/${event._id}`} style={{ marginRight: '10px' }}>       
+                                    <FontAwesomeIcon className='fa-eye' icon={faEye} />
+                                </Link>
+                                <Link to={`/edit-event/${event._id}`} style={{ marginRight: '10px' }}>       
+                                    <FontAwesomeIcon className='fa-edit' icon={faEdit} />
+                                </Link>
+                                <FontAwesomeIcon className='fa-delete' icon={faTrashAlt} onClick={() => handleDeleteEvent(event._id)}/>
+                            </div>
                         </div>
-
                     ))
                 ) : (
                     <p>No events hosted.</p>
                 )}
             </div>
             <div className='hosted-events'>
+                <h1>Your Posts:</h1>
+                {console.log(userPosts)}
                 {userPosts ? (
                     userPosts.map(post => (
-                        <div className='user-event-row'>
-                            <div className='buttons' style={{ display: 'flex', alignItems: 'center', marginBottom: '10px' }}>
-                                <div style={{ marginRight: '10px'}}>{post.title}</div>
-                                <button onClick={() => handleDeletePost(post.postId)}>Delete</button>
-                            </div>
+                        <div className='hosted-event'>
+                            <p>{post.title}</p>
+                            <FontAwesomeIcon className='fa-delete' icon={faTrashAlt} onClick={() => handleDeletePost(post.postId)}/>
                         </div>
-
                     ))
                 ) : (
                     <p>No posts created.</p>
                 )}
             </div>
-            <Link to={`/create-event`}>Create Event</Link>
-            <Link to={`/create-post`}>Create Post</Link>
-            <Link to={`/create-org`}>Create Org</Link>
-            <button onClick={() => handleSignOut()}>Sign Out</button>
+            <div className='profile-button-container'>
+                <Link to={`/create-event`} className='profile-page-button'>Create Event</Link>
+                <Link to={`/create-post`} className='profile-page-button'>Create Post</Link>
+                <Link to={`/create-org`} className='profile-page-button'>Create Org</Link>
+            </div>
+            <div className='profile-sign-out' onClick={() => handleSignOut()}>Sign Out</div>
         </div>
     )
 }
